This project is a Python script that generates strong and secure passwords based on user-defined criteria. The user can specify the length of the password and choose whether to include uppercase letters, digits, and special characters.
This project is a Python script that generates strong and secure passwords based on user-defined criteria.

------> Features
Customizable Length: Allows the user to specify the desired length of the generated password.

Character Type Selection: Lets the user choose whether to include uppercase letters, digits, and special characters in the password.

Secure Passwords: Generates strong and secure passwords by combining different types of characters.

User-Friendly: Simple and interactive script that prompts the user for their preferences and generates a password accordingly.

-------> How It Works
User Preferences: The script starts by asking the user for their preferences, including the desired password length and whether to include uppercase letters, digits, and special characters.

Character Pool Creation: Based on the user's choices, the script creates a pool of possible characters to be used in the password.

Password Generation: The script then randomly selects characters from the pool to generate a password of the specified length.

Display Password: The generated password is displayed to the user.