This project is a Python implementation of the classic two-player game Tic-Tac-Toe. Players take turns marking Xs and Os on a 3x3 grid, aiming to get three of their marks in a row, either horizontally, vertically, or diagonally.
This project is a Python implementation of the classic two-player game Tic-Tac-Toe. It provides an interactive and enjoyable way to play this timeless game on a computer.

-------> Features
Two-Player Mode: Allows two players to play against each other on the same device.

Interactive Gameplay: Players take turns to enter their moves and the game updates the board accordingly.

Win Detection: The game checks for a win condition after each move, detecting if a player has won by forming a line of three marks horizontally, vertically, or diagonally.

Draw Detection: The game also detects if the board is full without any player winning, declaring a draw.

Replay Option: After the game ends, players are prompted if they want to play again.

------> How It Works
Initialize the Board: The game starts with an empty 3x3 grid.

Player Turns: Players alternate turns to place their mark (X or O) on the board.

Check Win Conditions: After each move, the game checks if the current player has won by forming a line of three marks.

Check Draw Condition: The game also checks if the board is full and no player has won, resulting in a draw.

Display Result: The game displays the result, either declaring a win for one of the players or a draw.

Replay Option: Players are asked if they want to play another game after the current game ends.